refactor writings admin
