--- core/templates/site/writings/adminUserRolesPage.gohtml
@@ -1,44 +1,0 @@
-{{ template "head" $ }}
-<table border="1">
-    <tr>
-        <th>ID</th>
-        <th>User</th>
-        <th>Email</th>
-        <th>Roles</th>
-        <th>Actions</th>
-    </tr>
-    {{- range .Users }}
-    <tr>
-        <td>{{ .ID }}</td>
-        <td>{{ .Username.String }}</td>
-        <td>{{ .Email }}</td>
-        <td>
-            {{- if .Roles }}
-            <ul>
-                {{- range .Roles }}
-                <li>{{ .Name }}
-                    <form method="post" style="display:inline">
-        {{ csrfField }}
-                        <input type="hidden" name="permid" value="{{ .PermID }}">
-                        <button type="submit" name="task" value="remove">Remove Role</button>
-                    </form>
-                </li>
-                {{- end }}
-            </ul>
-            {{- else }}None{{- end }}
-        </td>
-        <td>
-            <form method="post">
-        {{ csrfField }}
-                <input type="hidden" name="username" value="{{ .Username.String }}">
-                <select name="role">
-                    {{- range $.Roles }}<option value="{{.Name}}">{{.Name}}</option>{{- end }}
-                </select>
-                <button type="submit" name="task" value="allow">Add Role</button>
-            </form>
-        </td>
-    </tr>
-    {{- end }}
-</table>
-<p>Permissions should be valid only.</p>
-{{ template "tail" $ }}
