+++ handlers/writings/routes_admin.go
@@ -9,9 +9,8 @@ import (
 // RegisterAdminRoutes attaches writings admin endpoints to ar.
 func RegisterAdminRoutes(ar *mux.Router) {
 	war := ar.PathPrefix("/writings").Subrouter()
-	war.HandleFunc("/users/roles", AdminUserRolesPage).Methods("GET")
-	war.HandleFunc("/users/roles", handlers.TaskHandler(userAllowTask)).Methods("POST").MatcherFunc(userAllowTask.Matcher())
-	war.HandleFunc("/users/roles", handlers.TaskHandler(userDisallowTask)).Methods("POST").MatcherFunc(userDisallowTask.Matcher())
+	war.HandleFunc("", AdminWritingsPage).Methods("GET")
+	war.HandleFunc("/", AdminWritingsPage).Methods("GET")
 	war.HandleFunc("/categories", AdminCategoriesPage).Methods("GET")
 	war.HandleFunc("/categories", handlers.TaskHandler(writingCategoryCreateTask)).Methods("POST").MatcherFunc(writingCategoryCreateTask.Matcher())
 	war.HandleFunc("/categories/category/{category}", AdminCategoryPage).Methods("GET")
